refactor(chart-card): rafactor code
