+++ OceanComponents/Classes/Extensions/PieChartDataSet+Extensions.swift
@@ -1,0 +1,17 @@
+//
+//  PieChartDataSet+Extensions.swift
+//  DGCharts
+//
+//  Created by Acassio Mendonça on 05/02/25.
+//
+
+import Foundation
+import DGCharts
+
+extension PieChartDataSet {
+    public func configureDataSetAppearance() {
+        drawValuesEnabled = false
+        drawIconsEnabled = false
+        selectionShift = 0.0
+    }
+}
